Fix servo initialization and only shift while driving
Fix servo initialization. The steering servo requires an additional
setPosition call (on top of the call in the setup function) to prevent
fully steering to the left until the synchronization with the
transmitter is established.

Only shift while driving to ensure quick engagement, lowering the stress
on the gearbox servo.