Improve and refactor HMI (incl Nextion display)
Implement the configuration of the HMI. The center and extreme positions
of the sticks can be teached via the menu on the Nextion display. These
settings are stored in the non-volatile memory (EEPROM) of the HMI
itself.

Improve high beam handling. The high beam is automatically switched of
when the low beam is off, similar to real cars.

Refactor the Teensy firmware, focusing on self-documenting code instead of
Doxygen comments.

Use 115200 baud for communication with Nextion HMI