Fix bugs sending and saving the configuation
Fix bugs in sending the configuration from the HMI to the VCU and saving
the configuration in the EEPROM of the VCU.

Handle the transmission as a two speed transmission (low and high, no
neutral). Initialise the transmission on the average of low and high and
move to either low or high as soon as the car starts driving. This
prevents putting too much strain on the transmission servo if the car is
switched off in second gear and cannot easily switch to first gear.